:recycle: - wip refactor Table
